Import ksquares-18.12.0; Tweak/OK bcallah@
Comment:
connect the dots to create squares game

Description:
KSquares is modeled after the well known pen and paper based game of
Dots and Boxes. The idea behind KSquares is to construct squares by
interconnecting the dots on the grid. Up to four players can play this
game together, taking turns. The trick is regardless of who has
constructed most of the square, it is the player who completes the
square that gets the point for it. To win the game the player has to
collect the most points by covering the most area with the squares of
player's own color.

Maintainer: Rafael Sadowski <rsadowski@openbsd.org>

WWW: https://www.kde.org/applications/games/ksquares