Import ksnakeduel-18.12.0; New DESCR and OK bcallah@
Comment:
Tron-like game

Description:
KSnakeDuel is a simple snake duel game for the KDE. You can play
KSnakeDuel against the computer or a friend. The aim of the game is to
live longer than your opponent. To do that, avoid running into a wall,
your own tail and that of your opponent.

KSnakeDuel also comes with KSnake, a simple Snake-like game for the KDE.
The aim of KSnake is to survive as long as possible and eat as many
fruits as you can.

You can switch between these games using the game type selector in the
configuration dialog.

Maintainer: Rafael Sadowski <rsadowski@openbsd.org>

WWW: https://www.kde.org/applications/games/ksnakeduel