import ports/fonts/tagmukay, a font used for Berber languages from George Rosamond, ok bentley@
Tagmukay is a Shifinagh script font with support for the Tawallammat
Tamajaq language. The script name is more commonly spelled Tifinagh,
but Shifinagh is the preferred spelling in the region where Tawallammat
Tamajaq is spoken. Tagmukay is a Unicode font that features
bi-consonant ligatures and alternate forms necessary to support
this language.

This port contains both TrueType and Web fonts.