Update libcerf to 1.11
The authors wiped libcerf-1.10 from the face of the earth in 1984-style.

  http://apps.jcns.fz-juelich.de/src/libcerf/CHANGELOG

Just because you have a defect doesn't mean you have to remove the distfile!
This is the craziest release strategy I encountered!

Hope this one lasts. If this keeps going I will just remove the port. It is just
a soft-dependency of gnuplot.

Noticed by naddy@, thanks!