Import frei0r-plugins-1.6.1; OK bcallah@
Comment:
minimalistic plugin API for video effects

Description:
Frei0r is a minimalistic plugin API for video sources and filters. The
behaviour of the effects can be controlled from the host by simple
parameters. The intent is to solve the recurring reimplementation or
adaptation issue of standard effects.

It is not meant as a generic API for all kinds of video applications.

There is no support for the requirements of special application areas
like non linear editors, hardware accelerated shader effects, and high
precision video processing. These advanced issues are not even solved
satisfactory for non cross application plugin apis and are still an
evolving field.

The frei0r API is not meant to be a competing standard to more
ambitious efforts.

Maintainer: Raphael Graf <r@undefined.ch>

WWW: https://frei0r.dyne.org/