unbreak some glib2 users on 32-bit arches, ok aja@
The old autoconf build has the following additional check after
determining a valid type for gsize, which is not present in meson.build.:

dnl If int/long are the same size, we see which one produces
dnl warnings when used in the location as size_t. (This matters
dnl on AIX with xlc)
dnl
AS_IF([test $ac_cv_sizeof_size_t = $ac_cv_sizeof_int &&
       test $ac_cv_sizeof_size_t = $ac_cv_sizeof_long],

It matters on OpenBSD too. Since my meson-fu is lacking, the patch
simply reorders the if statement to put the one we need earlier.