update lang/rust bootstraps to change a bit the way hash metadata is generated.
in general case, the output of `rustc -vV` is still used as part of the
hashed data. the output contains the exact version of the rustc and the
LLVM version used.

but when used for build the port, it will use a less changing string:
only the major+minor version (without patch number) and no more LLVM
version.

the choice between the two way is based on the presence of environment
variable.

it will permit to avoid change in filenames (and so PLIST regeneration
requirement) when:
- patch release of rustc is pushed
- llvm is updated