update to 72.0.3626.81
now chromium uses f'in java to use the closure compiler which
creates "better javascript" by parsing and converting all js
code into unreadable garbage. this feature is only enable on
amd64 because on i386 java randomly fails to allocate the needed
memory and on arm64 we do not have jdk yet. on these architectures,
we use the best method to create "better javascript", which is cat(1)