Import games/1oom, a game engine recreation of Master of Orion 1. Original submission by Max Fillinger, who takes MAINTAINER -- thanks! ok thfr@
1oom is a free and open source recreation of Master of Orion 1, a
classic turn-based strategy game from 1993 developed by Simtex and
published by Microprose. The player's goal is to expand their space
empire and ultimately rule the galaxy. Of course, rival empires have the
same idea...

While 1oom is free software, the game and its assets are not. You will
need to purchase a copy of Master of Orion 1 to use this package.