import exploitdb-papers
The Exploit Database is an archive of public exploits and corresponding
vulnerable software, developed for use by penetration testers and
vulnerability researchers. Its aim is to serve as the most comprehensive
collection of exploits, shellcode and papers gathered through direct
submissions, mailing lists, and other public sources, and present them
in a freely-available and easy-to-navigate database. This is the
papers section of it.

jasper@ suggested to put it under the books CATEGORY and generally fine with it
OK gonzalo@