Update to exiv2-0.27.0
Switch to cmake, enable i18n support and unit tests.

Also:
- don't install libxmp.a from the Adobe XMP SDK, to avoid confusion with
audio/libxmp
- don't provide the new Exiv2::getProcessPath() function.  Not portable,
doesn't work on OpenBSD, and shouldn't be exposed in the API of
a graphics library.  Instead, define a macro with CMake to help runtime
find gettext files.

Sample executables are now built and installed, not sure if all of them
are very useful.

From danj@ with input from sthen@, ajacoutot@ and me, tested in a bulk
by ajacoutot@