Import chivo-1.007.
Chivo (Goat) is a new Omnibus-Type grotesque Sans Serif typeface
family. The strength of Chivo Black makes it ideal for highlights
and headlines. Chivo Regular's elegance makes it ideal for combining
with the strength of Chivo Black and for for continuous reading
settings. Its design details make it an indispensable ally for any
designer.

From George Rosamond; thanks!

ok bcallah@