handle jdk 11 for unifi;
unifi requires javax.activation from activation.jar, which is present
and active in jdk 8, present but deprecated in 9/10 (hidden behind a
command-line flag), and removed in jdk 11.

add a FETCH_MANUALLY distfile to take this from Java Activation Framework;
this is under Oracle's binary license with an indemnity clause so you get
to download it yourself, but it's fairly straightforward. adapt the startup
script to cope.