Import fonts/taviraj, a Serif Latin and Looped Thai typeface. Original submission by George Rosamond, who takes MAINTAINER -- thanks! Tweaks and ok sthen@
The objective of this project is to create a formal Serif Latin &
Looped Thai typeface. Taviraj is a contemporary serif typeface
family that maintains readability and legibility. Inspired by
transitional serif typeface design, Taviraj has 9 weights and their
italics, which range from Thin to black. It is a contemporary take
on high contrast typefaces that never go out of style.