Fix py-matplotlib-2.2.4 build.
naddy@ and ajacoutot@, independently from each other, found that the
recent update of py-matplotlib does not build. Issue is caused by me not
noticing that matplotlib tries to download jquery-ui-1.12.1, which fails
when building in a chroot using dpb. Instead of downloading jquery-ui
use the bundled files.

REVISION is not bumped as patched file is not included in the package,
and no new files are added.