Import games/gnukem, an open source retro-style 2D scrolling platform shooter game. Big-endian fixes from Raphael Graf. macppc testing and ok cwen@
Dave Gnukem is an open source retro-style 2D scrolling platform shooter,
inspired by and similar to Duke Nukem 1 (a famous original 1991 game
that launched the Duke Nukem series). It features an integrated level
editor. (Please note this is not a 'clone' or a 're-make'; it is
probably most akin to a parody.)