Import public-sans-1.0.0.
Public Sans is a strong, neutral, principles-driven, open-source typeface
for text or display. It is a fork of the SIL Open Licensed face Libre
Franklin. Public Sans has many similarities with its parent, but differs
in its focus on longform reading and neutral UI applicability. It takes
inspiration from geometric sans faces of the 20th century, as well as the
original Franklins of the 19th, resulting in something of a mongrel face
that retains its American origin.

ok sthen@