Update to borgbackup-1.1.10.
Bugfix release. Noticeable change is that the latest supported
msgpack-python release is bundled. Changelog can be found at
https://github.com/borgbackup/borg/blob/1.1.10/docs/changes.rst#version-1110-2019-05-16

Switching to the bundled version of msgpack-python makes sense as the
version in ports is in need of an update: msgpack-python has been
renamed by its upstream, and borgbackup was seemingly the only port
unable to cope with this renaming, thus blocking an update.

OK kn@