import steamworks-nosteam, a stubbed C# interface to the Steam API, based on MIT-licensed full interface library. ok bcallah@ (thanks for spotting the typo!) ok kn@
DESCR:
Drop-in replacement for Steamworks.NET.dll to run without Steam
client present, e.g. if running on an operating system that doesn't
feature a Steam client. A number of games are made largely
platform-independent, but use Steamworks.NET and error out if no
Steam client is present. This project builds a .dll that can be
dropped in and is intended to keep the parent program running even
in the absence of a Steam client.