phase out old PERMIT_* variables. We won't be distributing CDROM anymore, so simplify to just PERMIT_PACKAGE / PERMIT_DISTFILES
In particular, the new variables are shorter, so this makes
for better cosmetic sense in ports.

The "current" version allows for the old variables to die out
peacefully, at our leasure, and then I'll remove the old stubs.

As discussed with sthen@, various people agree, and deraadt@
is okay with the strategic change.

This does survive a bulk.

PLEASE NOTE: built packages require current pkg* tools.
In particular, the current version deals with
@comment pkgpath=* ftp=*
just fine, but the old one WILL COMPLAIN about missing cdrom info.

register-plist also doesn't care, adding/removing cdrom info
is a no-op for it. There is NO BUMP needed for the conversion.

For now, sqlports STILL carries the old variables. I'll deal
with their removal later.