import net/dnscontrol, from Karlis Mikelsons, much tidying and ok from kn@
DNSControl is a system for maintaining DNS zones. It has two parts: a
domain specific language (DSL) for describing DNS zones plus software
that processes the DSL and pushes the resulting zones to DNS providers
such as Route53, CloudFlare, and Gandi. It can talk to Microsoft
ActiveDirectory and it generates the most beautiful BIND zone files
ever.