import glslang for Vulkan API support; ok jsg@
DESCR:
An OpenGL and OpenGL ES shader front end and validator.

There are several components:

A GLSL/ESSL front-end for reference validation and translation of
GLSL/ESSL into an AST.

An HLSL front-end for translation of a broad generic HLL into the AST.

A SPIR-V back end for translating the AST to SPIR-V.

A standalone wrapper, glslangValidator, that can be used as a
command-line tool for the above.