Update audio/picard 0.16 -> 2.1.3
The project no longer bundles the plugins with the source. So I got
rid of all the Makefile logic for the plugins.

Port moved to using MODPY_PI. There is a patch for setup.py to use
the setuptools version of install which has support for
--single-version-externally-managed. I've sent this patch upstream
and it has already been committed so will be in the next release.
I discarded the old setup.py patch.

A chunk of tests fail because the PyPI tarball id missing any .aac
files from test/data. I got a patch committed upstream for their
MANIFEST.in file so that future releases will include those files.

Dependencies on ffmpeg, liboaf, and others were dropped as they
aren't needed for recent releases of picard.

Input from sthen@

ok edd@ sthen@