Import games/openjazz, an open-source version of the classic Jazz Jackrabbit games. macppc testing and ok cwen@
OpenJazz is a free, open-source version of the classic Jazz Jackrabbit
games. Jazz Jackrabbit is a platforming game in which the player
controls an anthropomorphic rabbit against an invading tortoise army.

While this package and the game engine are open source, one still needs
to provide the original game assets from Jazz Jackrabbit, which is not
open source. You will need to purchase the original game to use this
package.