Import py-scikit-learn.
scikit-learn is a BSD-licensed Python module for machine learning
and data mining built on top of SciPy.

Initially created by robert@, modified by me and kmos@.

This is not polished enough, but it can improved in the tree now.
Once the 21 branch rolls out a new version, we will probably update to that.
Currently it is broken for non-OpenMP builds.

OK robert@, kmos@