Update textproc/py-xmldiff 0.6.10 -> 2.3
py-xmldiff hasn't been updated in a long time, but it still gets used.
So time to update it. Current 2.x versions are a rewrite in pure python.

Patches go away. Port moved to use MODPY_PI and MODPY_SETUPTOOLS.

Move it to python3 only since executable is just xmldiff and xmlpatch.

OK landry@