add an electronjs (https://electronjs.org/) port;
the electronjs port is going to live inside the chromium port,
as a separate flavor, which is called electron (obviously)
because it requires all of our chromium patches as well and
this way we can avoid having to maintain another subdirectory
with hundreds of patches; this is mostly unconventional but
for me this is the best solution (maintainence-wise)

a couple of patches for the actual electron source are shamelessly
stolen from freebsd

this commit also adds a --no-sandbox flag because electron
cannot use unveil and pledge just yet, i will work on that
in the future

basic stuff works, but don't expect anything fancy yet

--

initial versoin of electron-6.0.3:
Build cross platform desktop apps with JavaScript, HTML, and CSS.

It's easier than you think.

If you can build a website, you can build a desktop app. Electron is a
framework for creating native applications with web technologies like
JavaScript, HTML, and CSS. It takes care of the hard parts so you can
focus on the core of your application.