Import Accerciser 3.33.4
Accerciser is an interactive Python accessibility explorer for the GNOME
desktop. It uses AT-SPI2 to inspect and control widgets, allowing you to
check if an application is providing correct information to assistive
technologies and automated test frameworks. Accerciser has a simple
plugin framework which you can use to create custom views of
accessibility information.

ok ajacoutot@