Ghidra's build system prefers g++ over clang, which is not what we want on a clang-based platform like amd64.  This commit makes Ghidra build with clang by:
* Adding a clang toolChains block in nativeBuildProperties.gradle
* Adding COMPILER="base-clang ports-clang"
* Updating WANTLIB by replacing libstdc++ with ${COMPILER_LIBCXX}

In addition, honor CXX by setting tools.cppCompiler.executable and
tools.linker.executable in nativeBuildProperties.gradle to ${CXX}.  Also print
those variables so that we can confirm that they are set properly.

Lastly, add --stacktrace to the gradle command to facilitate future debugging.

Thanks to daniel@ for bringing this to my attention, naddy@ for his detailed
report of the g++ and CXX issues, and sthen@ for suggesting a possible fix.

ok naddy@ sthen@