Let print/poppler *build* depend on devel/boost.
When boost is detected during configure time, poppler uses small_vector,
which is a header-only class of boost. This can lead to build
failures if boost is removed (e.g. by dpb) shortly after the build
of poppler starts, as noticed by naddy@.

This doesn't restrict the arches where poppler can be built, because
it c++11 is needed anyway (noted by sthen@).