import polybar-3.4.0
Polybar is a fast and easy to use tool for creating status bars.
Polybar aims to help users build beautiful and highly customizable
status bars for their desktop environment, without the need of having a
black belt in shell scripting.

currently not all the OS-specific modules support OpenBSD (cpu does work now).
the remainder of modules will be worked on