Fix missing Okular WANTLIB, turn kinit from LIB into RUN_DEPENDS
Okular5Core is dlopen(3)ed at runtime, thus port-lib-depends-check does not
pick it up.  This causes okular to be omitted as RUN_DEPENDS and starting
kile on systems without it would exit immediatly with a warning box.

kinit's shared library is not used at all, so make that clear while here.

OK rsadowski