Update dpic to 2019.08.30. This fixes CVE-2019-13989.
The codebase is now fully in C. As a result, we can remove previous
patches, they are not needed anymore as p2c is gone, and upstream now
has the sprintf -> snprintf changes in wfloat.

The remaining sprintf -> snprintf changes have been moved to a new
patch, as those functions are now in their own file (pdf.c).

OK bentley@