Remove postgresql-plv8
The version in-tree is very old and requires significant work to
support new major versions of PostgreSQL.  The newer versions use
much different build system that I wasn't able to get working.

Anyone is welcome to readd this if they are able to get it working
and are willing to maintain it.

OK pea@