Import x11/kde-applications/kturtle (ok sthen@, Tested in a  bulk, incredibly many thanks!)
Is is an one by one replacement for x11/kde4.

KTurtle is an educational programming environment for the KDE Desktop. KTurtle
aims to make programming as easy and touchable as possible, and therefore can
be used to teach kids the basics of math, geometry and... programming.

The programming language used in KTurtle is loosely based on Logo.  KTurtle
allows, just like some implementations of Logo, to translate the programming
language (the commands, the documentation and the error messages) to the native
language of the programmer. For more information on translating KTurtle please
read the translation how to. Translating the programming language to the native
language of the programmer is one of the many ways KTurtle tries to make
learning to programming more simple. Other features to help to achieve this
goal are: intuitive syntax highlighting, simple error messages, integrated
canvas to make drawings on, integrated help function, slow-motion or step
execution, and more.