Import x11/kde-applications/rocs (ok sthen@, Tested in a  bulk, incredibly many thanks!)
Is is an one by one replacement for x11/kde4.

Rocs aims to be a Graph Theory IDE for helping professors to show the results
of a graph algorithm and also helping students to do the algorithms. Rocs has a
scripting module, done in Qt Script, that interacts with the drawn graph and
every change in the graph with the script is reflected on the drawn one.