Update print/texlive to version 2019.
A few notes:

This year, we install the interpreter symlinks when the supporting texmf
files become available. Before we installed them all in texlive_base and
you could get dangling symlinks and confused users.

There are also a lot of Ruby-1.8 scripts in TeX Live, but we no longer
package such an old Ruby. There's too much to systematically test, so
for now we re-write the #! to a newer Ruby, and we can fix any problems
that arise on a case-by-case basis.

Tested by: matthieu@, cwen@, Jon Bernard and jca@.

Additional input from sthen@ and cwen@.

OK cwen@ and jca@.

Thank you all for your help!